file structure set up.
