+++ src/App.css
@@ -1,6 +1,6 @@
-/* .App {
+.App {
   text-align: center;
-} */
+}
 
 .App-logo {
   animation: App-logo-spin infinite 20s linear;
@@ -9,13 +9,9 @@
 
 .App-header {
   background-color: #222;
+  height: 150px;
   padding: 20px;
   color: white;
-  text-align: center;
-}
-
-.App-title {
-  font-size: 1.5em;
 }
 
 .App-intro {
