+++ src/components/Footer/Footer.js
@@ -1,0 +1,11 @@
+import React from "react";
+import "./Footer.css";
+
+const Footer = () => (
+  <div className="footer">
+    <p>Footer goes here</p>
+
+  </div>
+);
+
+export default Footer;
