+++ src/components/Main/Main.js
@@ -1,0 +1,17 @@
+import React from "react";
+import "./Main.css";
+
+
+const Main = () => (
+ <div classNameName="main">
+   
+
+   <div className="card">
+   <p>Main content goes here</p>
+
+   </div>
+   </div>
+    );
+    
+
+export default Main;
