+++ src/components/Navbar/Navbar.css
@@ -1,0 +1,4 @@
+.jumbotron {
+  background-color: lightgray;
+  border-bottom: 1px solid black;
+}
