+++ src/components/Sidebar/Sidebar.js
@@ -1,0 +1,13 @@
+import React from "react";
+import "./Sidebar.css";
+
+const Sidebar = () => (
+ <div classNameName="Sidebar">
+   
+<p id="side-bar">Sidebar goes here</p>
+  
+</div>
+);
+
+
+export default Sidebar;
